Update libVA/driver to support frame skip feature which is important for WiDi clone mode power-saving
BZ: 156748

libVA and GEN Driver need to be upgraded to support Frame Skip Feature.

This feature is important for WiDi clone mode power saving. Currently there is a big gap for power consumption (about 40%) for WiDi clone mode on VLV2-A

This change is a simple squash merge from main branch

Signed-off-by: Austin Yuan <shengquan.yuan@intel.com>
Change-Id: I4163a4f05cd9025728cfaf3b44a38f82cf4ff526