Do not hardcode LIBVA_DRIVERS_PATH by constant value
In 64bit, there's such chance that we will use 32bit libva.so when mediaserver is
built to 32bit. So we cannot hardcode LIBVA_DRIVERS_PATH as a constant value at build
time. We need to let the compiler decide whether to use /system/lib or /system/lib64
during build time

Change-Id: I3dff0099f9ae42d683d489d601581e8487915e57
Signed-off-by: Qiming Shi <qiming.shi@intel.com>